LogginService/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.1
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kafka-python==2.0.2
MarkupSafe==2.1.1
mongoengine==0.24.1
pydantic==1.9.0
pymongo==4.0.2
python-dotenv==0.20.0
python-on-whales==0.41.0
requests==2.27.1
tqdm==4.63.1
typer==0.4.1
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.7.0
